Streamlit/requirements.txt
firecrawl-py
openai
python-dotenv
requests
streamlit
youtube-transcript-api